+++ FarmingTracker/Stat/Profit.cs
@@ -4,22 +4,22 @@ namespace FarmingTracker
 {
     public class Profit
     {
-        public int MaxProfitInCopper { get; private set; }
-        public int MaxTpProfitInCopper { get; private set; }
-        public int TpSellProfitInCopper { get; private set; } // sell by listing a sell order
-        public int TpBuyProfitInCopper { get; private set; } // sell by instant sell to buy order
-        public int VendorProfitInCopper { get; private set; } // sell to npc vendor
+        public long MaxProfitInCopper { get; private set; }
+        public long MaxTpProfitInCopper { get; private set; }
+        public long TpSellProfitInCopper { get; private set; } // sell by listing a sell order
+        public long TpBuyProfitInCopper { get; private set; } // sell by instant sell to buy order
+        public long VendorProfitInCopper { get; private set; } // sell to npc vendor
         public bool CanNotBeSold => !CanBeSoldToVendor && !CanBeSoldOnTp;
         public bool CanBeSoldToVendor { get; set; }
         public bool CanBeSoldOnTp => TpSellProfitInCopper > 0 || TpBuyProfitInCopper > 0;
 
-        public void SetVendorProfit(int vendorProfitInCopper)
+        public void SetVendorProfit(long vendorProfitInCopper)
         {
             VendorProfitInCopper = Math.Abs(vendorProfitInCopper);
             SetMaxProfit();
         }
 
-        public void SetTpSellAndBuyProfit(int tpSellPriceInCopper, int tpBuyPriceInCopper)
+        public void SetTpSellAndBuyProfit(long tpSellPriceInCopper, long tpBuyPriceInCopper)
         {
             // 85/100 is -15% tp fee with integer rounding
             TpSellProfitInCopper = Math.Abs(tpSellPriceInCopper) * 85 / 100;
